# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Law Firm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lawyer"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lawyers:
    get:
      operationId: getLawyers
      tags: [lawyers]
      summary: List all lawyers
      responses:
        "200":
          description: Lawyers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lawyer"
    post:
      operationId: createLawyer
      tags: [lawyers]
      summary: Create a lawyer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLawyerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lawyer"
  /lawyers/{id}:
    get:
      operationId: getLawyer
      tags: [lawyers]
      summary: Get a lawyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lawyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lawyer"
    put:
      operationId: updateLawyer
      tags: [lawyers]
      summary: Update a lawyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLawyerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lawyer"
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Clients list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientsPage"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cases:
    get:
      operationId: getCases
      tags: [cases]
      summary: List all cases
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, closed, archived]
        - name: lawyerId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Cases list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasesPage"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case with timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case with timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cases/{id}/timeline:
    post:
      operationId: addTimelineEvent
      tags: [cases]
      summary: Add timeline event to a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        lawyer:
          $ref: "#/components/schemas/Lawyer"
        message:
          type: string
      required: [lawyer]
    Lawyer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        oab:
          type: string
        specialty:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        isAdmin:
          type: boolean
        activeCaseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, oab]
    CreateLawyerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        oab:
          type: string
        specialty:
          type: string
        bio:
          type: string
        isAdmin:
          type: boolean
      required: [name, email, password, oab]
    UpdateLawyerBody:
      type: object
      properties:
        name:
          type: string
        oab:
          type: string
        specialty:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        cpf:
          type: string
        address:
          type: string
        notes:
          type: string
        activeCaseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        cpf:
          type: string
        address:
          type: string
        notes:
          type: string
      required: [name, email]
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        cpf:
          type: string
        address:
          type: string
        notes:
          type: string
    ClientsPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    Case:
      type: object
      properties:
        id:
          type: integer
        caseNumber:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, pending, closed, archived]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        caseType:
          type: string
        clientId:
          type: integer
        lawyerId:
          type: integer
        client:
          $ref: "#/components/schemas/Client"
        lawyer:
          $ref: "#/components/schemas/Lawyer"
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, priority, clientId, lawyerId]
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/Case"
        - type: object
          properties:
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
    CreateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, pending, closed, archived]
          default: pending
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        caseType:
          type: string
        clientId:
          type: integer
        lawyerId:
          type: integer
        deadline:
          type: string
          format: date-time
      required: [title, clientId, lawyerId]
    UpdateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, pending, closed, archived]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        caseType:
          type: string
        lawyerId:
          type: integer
        deadline:
          type: string
          format: date-time
    CasesPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Case"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        lawyerId:
          type: integer
        lawyer:
          $ref: "#/components/schemas/Lawyer"
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [note, hearing, document, deadline, update, status_change]
        eventDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, caseId, title, eventType, eventDate]
    CreateTimelineEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [note, hearing, document, deadline, update, status_change]
          default: note
        eventDate:
          type: string
          format: date-time
      required: [title, eventType, eventDate]
...